Change: map display to its own component
